Validacion de fechas para la rendicion hecha

+++ Modelo/Rendicion.cs
@@ -12,19 +12,19 @@ namespace PagoAgilFrba.Modelo
         private DateTime fechaRendicion;
         private int totalRendicion;
         private int porcentajeComision;
-        private double importeComision;
+        private float importeComision;
         private int cantFacturasRendidas;
 
         public void setFechaRendicion(DateTime unaFecha)    { this.fechaRendicion = unaFecha; }
         public void setTotalRendicion(int unTotal)          { this.totalRendicion = unTotal; }
         public void setPorcentajeComision(int porc)         { this.porcentajeComision = porc; }
-        public void setImporteComision(double importe)      { this.importeComision = importe; }
+        public void setImporteComision(float importe)       { this.importeComision = importe; }
         public void setCantFacturasRendidas(int cant)       { this.cantFacturasRendidas = cant; }
 
         public DateTime getFechaRendicion()     { return this.fechaRendicion; }
         public int getTotalRendicion()          { return this.totalRendicion; }
         public int getPorcentajeComision()      { return this.porcentajeComision; }
-        public double getImporteComision()      { return this.importeComision; }
+        public float getImporteComision()       { return this.importeComision; }
         public int getCantFactRendidas()        { return this.cantFacturasRendidas; }
